Updated Forestry API

+++ src/minecraft/forestry/api/apiculture/IBeeMutation.java
@@ -5,5 +5,14 @@ import forestry.api.genetics.IGenome;
 import forestry.api.genetics.IMutation;
 
 public interface IBeeMutation extends IMutation {
-	int getChance(IBeeHousing housing, IAllele allele0, IAllele allele1, IGenome genome0, IGenome genome1);
+	
+	/**
+	 * @param housing
+	 * @param allele0
+	 * @param allele1
+	 * @param genome0
+	 * @param genome1
+	 * @return float representing the chance for mutation to occur. note that this is 0 - 100 based, since it was an integer previously!
+	 */
+	float getChance(IBeeHousing housing, IAllele allele0, IAllele allele1, IGenome genome0, IGenome genome1);
 }
